+++ src/main/java/racingCar/domain/Cars.java
@@ -1,4 +1,6 @@
-package racingCar;
+package racingCar.domain;
+
+import racingCar.domain.Car;
 
 import java.util.ArrayList;
 import java.util.List;
