+++ src/main/java/racingCar/view/InputView.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.view;
 
 import java.util.HashMap;
 import java.util.Map;
